LiquidWeb offers Managed servers also
LiquidWeb also offers Managed servers in addition to Shared servers.

With Managed Server [dedi or vps] the end user has much more freedom to configure the server as they'd like, unlike shared servers. So depending on the version of EasyApache they are using, on the server, they should expect access to PHP 5.3 - 7.0. There are no PHP 7.0 enabled Shared servers yet, but it's certainly offered on Managed servers.

If you're not familiar with EasyApache:
On EasyApache3 they can have a single version of 5.3.x - 5.6.x.
On EasyApache4 then can have either multiple or a single PHP version choosing between 5.4.x - 7.0.x.

Please let me know if this needs to be formatted any differently for the purposes of this site. Thanks!