PHP 7.4 support for Pantheon
Also, I have updated the patch version for all of our older releases in this table. It appears that this is perhaps unnecessary, as the [PHP Versions table](http://phpversions.info/paas-hosting/) has more up-to-date values than what is shown here. The odd thing, though, is that the PHP 7.2 column shows version 7.2.8, even though your crawler ran on us 17 hours ago, and [our PHP 7.2 php-info URL is reporting version 7.2.29]. All other PHP versions in the table are correct, though, so it's not really clear why just this one old version is stale.

🤷‍♂